requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Limiter
Flask-Talisman
Flask-Compress
Flask-SocketIO
Flask-Cors
Flask-Caching
Flasgger
gunicorn
eventlet
psycopg2-binary
requests
python-dotenv
Brotli
sendgrid
